+++ packages/js/src/RoomSessionScreenShare.ts
